requirements.txt
line-bot-sdk==2.2.1
flask==2.0.1
flask_caching==1.10.1
requests==2.27.1